neural-networks/object-detection/barcode-detection-conveyor-belt/requirements.txt
depthai>=3.0.0
depthai-nodes==0.3.4
numpy>=1.22
opencv-python-headless~=4.10.0
pyzbar==0.1.9
Pillow==12.0.0